requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
requests==2.32.3
beautifulsoup4==4.13.3
python-dotenv==1.0.1

google-cloud-bigquery==3.29.0
